extract demo files
